pyproject.toml
[tool.poetry]
name = "cyrillicsegnet"
version = "0.1.0"
description = ""
authors = ["abdulrahmandawoud <dawoud.abdulrahman.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
pillow = "^11.0.0"
ipykernel = "^6.29.5"
keras = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
